pyproject.toml
[tool.poetry]
name = "awful-auth"
version = "0.1.0"
description = "Authentication service for Goon Auth Network"
authors = ["Ian P <ian@aotc.us>"]
repository = "https://github.com/GoonAuthNetwork/awful-auth"
license = "GPL-3.0"
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.8"
pydantic = "1.8.1"
fastapi = "0.63.0"
uvicorn = "0.13.4"
gunicorn = "20.0.4"
redis = "3.5.3"
requests = "2.25.1"
httpx = "0.17.1"
python-dotenv = "0.15.0"
fakeredis = "1.4.5"
loguru = "0.5.3"
beautifulsoup4 = "4.9.3"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.0"
black = "^20.8b1"
pytest = "^6.2.2"
pytest-httpx = "^0.11.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
